requirements.txt
django
python-dotenv
python-memcached
postgres
psycopg2-binary
gunicorn
django-sslserver
django-debug-toolbar
ldap3
django-url-or-relative-url-field

